Make it possible to poll every X seconds, instead of indefinitely
By introducing an options pullIntervalSeconds, users can opt to poll SQS
less frequently. For instance, if set to 60, we will wait 60 seconds after
we get a response from SQS before starting a new poll request.

If the option is not set, it should work as before, so this is an opt-in
feature for people who might care about the number of SQS requests they
use.